+++ R/ggsurv.R
@@ -51,7 +51,7 @@ if(getRversion() >= "2.15.1") {
 #'
 #'   # Adjusting the legend of the ggsurv fit
 #'   p_(pl.sex +
-#'     ggplot2::guides(linetype = FALSE) +
+#'     ggplot2::guides(linetype = "none") +
 #'     ggplot2::scale_colour_discrete(
 #'       name   = 'Sex',
 #'       breaks = c(1,2),
@@ -92,7 +92,7 @@ if(getRversion() >= "2.15.1") {
 #'         direction = 1
 #'       )(4)
 #'     ) +
-#'     ggplot2::guides(color = FALSE, linetype = FALSE))
+#'     ggplot2::guides(color = "none", linetype = "none"))
 #' }
 ggsurv <- function(
   s,